break-out SkipLoginButton
